pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "audiobot"
version = "0.1.0"
description = "Holy Spirit Vocal Engine: audio cleanup, stems, batch + web"
requires-python = ">=3.10"
authors = [{name = "JesusCartel"}]
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "jinja2",
    "numpy",
    "librosa",
    "soundfile",
    "pyloudnorm",
    "scipy",
    "noisereduce",
    "requests",
    "google-cloud-storage",
    "python-dotenv",
]

[project.scripts]
audiobot = "audiobot.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["audiobot*"]

[tool.setuptools.package-data]
"audiobot.web" = [
    "templates/*.html",
    "static/*",
    "verses_default.txt",
]
